Add main command with subcommands
This project is inspired by the `protokube` service that is
included with https://github.com/kubernetes/kops. Whereas
`protokube` works as a continuously running service in a
container, `keights` is a set of cli commands to be managed by
systemd using timers and path units.

The package contains three subcommands:

`collect`: This collects network information about an autoscaling
group and writes it to a file.

`volumize`: This attaches an EBS volume with a specified tag to
the EC2 instance the command is running on, and creates a
filesystem on the volume if it does not exist.

`share`: This writes entries to `/etc/hosts` so that hosts within
the cluster can find each other.