Make autoscaling group name autodiscoverable
Putting a reference to the autoscaling group name in the launch
configuration of a CloudFormation template causes a circular
reference, since the autoscaling group depends on the launch
configuration. The keights commands are being changed here to get
the name of the autoscaling group that they belong to at runtime
instead.