Add options to kubelet systemd unit for node labels
The option `--register-node=true` has been removed as a default
and is now replaced by the value from `KEIGHTS_NODE_REGISTER`
environment variable.

For a master, the option `KEIGHTS_NODE_REGISTER` should be set to
`--register-with-taints=kubernetes.io/role=master:NoSchedule`,
and for a node, it should be `--register-node=true`

To label a node: for a master, the option `KEIGHTS_NODE_LABEL`
should be
`kubernetes.io/role=master,node-role.kubernetes.io/master=`, and
for a node, it should be
`kubernetes.io/role=node,node-role.kubernetes.io/node=`.