Refactor to make CloudFormation responses more reusable
Add a function for creating a new response body with defaults
set from a handler's arguments was added to the `response`
package.

Remove `sendErrorResponse()` from `subnet_to_az` package and
replace with the more generic `FireResponse()` in the `response`
package. This function can fire a success or failure response,
but returns no value since the Lambda will exit after it has run.