Remove `KEIGHTS_ASG_NAME` environment variable from systemd unit
Because the environment variables for the keights-collector
service are configured through a launch configuration, putting the
name of the ASG into it creates a circular dependency. The
keights collect subcommand has already been updated to do a lookup
on the autoscaling group to get the name, and it needed to be
removed from the systemd unit.