Enable SSL
The apiserver now has SSL enabled, as well as SSL bootstrapping
of kubelets throughout the cluster.

The load balancer was changed to be a network load balancer, doing
a TCP pass-through to the apiserver, since the certificates are
signed by an internal CA. The CA is deployed as part of this stack
via a custom CloudFormation resource backed by a Lambda.