Fix controlPlaneEndpoint port
The controlPlaneEndpoint setting does not work with a port added.
The bind port of the apiserver is added to controlPlaneEndpoint.
This makes it necessary for the load balancer to listen on the
same port as the apiserver. This is unfortunate, but until that is
fixed, the load balancer will be changed to listen on 6443.