Rework security groups
Security group definitions are all put into the `common.yml`
CloudFormation template, so all the rules between them can be
done ahead and then applied to the other stacks.

Traffic has been opened between all nodes in the cluster, except
for nodes having access to etcd on the masters. Later this may be
updated to be more specific, but there is little documentation on
which ports are required to be opened.