Make S3 bucket for Lambda archive files a parameter
Deploying Lambdas from archives stored in S3 buckets requires
that the bucket live in the region where the Lambda is deployed.
Now the bucket is a parameter so that a user can copy the zip
files to their own bucket in their own region. By default, the
value is `cloudboss-public` in us-east-1.