Add resource signals to master CloudFormation stack
For doing stack creation and rolling updates of the masters, it
should be possible to do a health check and then signal the
stack when the apiserver is running before the next instance is
replaced.

A new command, `keights signal`, has been added which acts like
Amazon's `cfn-signal` command, so that the stack may be
signaled during creation or updates. The master stack
autoscaling group resource has a creation policy and update
policy added which will wait for the signals coming from
`keights signal`.

A new command was added to `keights` rather than using
`cfn-signal` because that would have required adding Python 2
to the AMI, and the implementation to do it was simple enough.