Modify subnet_to_az Lambda to output list of AZs from subnets
By outputting a list of availability zones given a list of subnet
IDs, the subnet_to_az Lambda only needs to be called once, and the
output for an individual AZ can be retrieved by using `Fn::Split`
in CloudFormation. The subnet_to_az function must output the AZs
in the same order as the subnet inputs.