Update masters to remove collect and share services
Collecting IPs and sharing via /etc/hosts is no longer necessary,
now that the `auto_namer` Lambda is in place to add DNS records.

This also includes having kubeadm generate certificates for etcd,
so that if the naming changes, the certs will be signed for the
new names as well. It also makes certificate rotation much easier,
as it will happen automatically on stack updates.