Update to run Kubernetes 1.11
Most of the changes for Kubernetes 1.11 focus on removing the
static pod manifest for etcd, and allowing kubeadm to handle that.
The kubeadm init config file has been expanded to specify the
options that should be passed to etcd, as well as updates to the
kubelet configuration. The kubelet configuration is mostly
handled by kubeadm now, so most of the options that were formerly
passed in the systemd dropin have been removed.