Improve resiliency of master initialization
This improves the volumize command:
The `WaitForVolume()` function adds `status: available` to its
filters when querying AWS. This should make the `AttachVolume`
operation more likely to succeed. Retry is also added around
`AttachVolume` so that it does not abort on the first error.
Length of time to attach and wait for volume is also increased.

The master.yml CloudFormation template has the timeout for
receiving a signal from masters increased to 30 minutes, for cases
where volumize is taking especially long to run.