Derive Kubernetes version from keights version in Concourse task
This changes the `build-ami` task to get the version of k8s from
the tag or semver used for the keights version, so the two are
always in sync, and a change to the keights repo is not needed to
build a new version of Kubernetes. Instead, the snapshot semver
can be bumped in the git repo where the semver is stored.