Use Lambda to disable source dest check on EC2 instances
When kube-router handles disabling of source dest check, it can
cause EC2 throttling if the cluster size gets large enough. This
ensures that the attribute is only set once on instance launch,
and disables kube-router from setting the attribute. This also
modifies calico to use `CrossSubnet` mode for IPIP, since the
Lambda can now set the attribute, whereas before it was unset when
using calico.