Add `ebsnvme-id` command with corresponding udev rules to AMI
The `ebsnvme-id` command is used in Amazon Linux AMIs to map nvme
device names to the name given to EC2 for an EBS volume device.
For example, if a volume is created with the device name
`/dev/xvdg`, it may have an internal nvme device name of, say,
`/dev/nvme1n1`, though this is not always predictable. The udev
rule runs `ebsnvme-id` to determine the name chosen by the user,
and then creates a symlink for it to the nvme device.

The command used by Amazon contains Python 2 syntax, so it was
modified to work with Python 3, as that is the default in Debian
stretch.