Remove use of etcd internal device
Since adding the ebsnvme-id script into the AMI with a udev rule
to create a symlink to the name given to Amazon for the etcd
block device mapping, there is no more internal name vs external
device name. This should no longer be surfaced to the user as
CloudFormation parameters, or as arguments to the keights binary.