Generate front proxy CA from kube_ca Lambda
API aggregator test failed with the following error due to kubeadm
signing separate front proxy CA certificates on each master:

      <*errors.StatusError | 0xc0021d7050>: {
          ErrStatus: {
              TypeMeta: {Kind: "", APIVersion: ""},
              ListMeta: {SelfLink: "", ResourceVersion: "", Continue: ""},
              Status: "Failure",
              Message: "the server has asked for the client to provide credentials",
              Reason: "Unauthorized",
              Details: {
                  Name: "",
                  Group: "",
                  Kind: "",
                  UID: "",
                  Causes: [
                      {
                          Type: "UnexpectedServerResponse",
                          Message: "unknown",
                          Field: "",
                      },
                  ],
                  RetryAfterSeconds: 0,
              },
              Code: 401,
          },
      }
      the server has asked for the client to provide credentials

This commit enhances the kube_ca Lambda to generate the front proxy
CA certificate and key to be shared among all masters, as is done
with the other CA certificates.

In order to support introducing a new SSM parameter to an existing
stack, the Lambda has been modified to handle CloudFormation stack
updates, which it previously ignored. Because of this, it also
runs the delete on a stack rollback. It is undesirable to delete
certificates and keys on a rollback, so the Lambda no longer does
deletion of SSM parameters.