Increase number of verify-network pods
When the conformance test starts running in upgrade-cluster Concourse job, there
are sometimes timeouts when connecting to the apiserver. This increases the
number of pods that run the verify-network test before starting the conformance
test, to ensure the network has really settled. There should be approximately
four passes of the job per node.