Update Ansible roles to support Ansible 2.9
Since modifying the autoscaling groups in the CloudFormation templates to use
LaunchTemplates instead of LaunchConfigurations, Ansible needs to be updated
in order to be able to run with state=absent. This is because in older versions
of Ansible, the ec2_asg_facts module assumed that autoscaling groups had a
LaunchConfiguration, and the absent.yml tasks uses that module.

Changes include:
* Updating all relevant requirements.txt files to use Ansible 2.9.x and
  other updated dependencies.
* Updating roles to replace *_facts modules with newer *_info modules.
* Replacing k8s_raw module with k8s module.
* Applying calico manifests in two parts to work with newer k8s module.