Mount etcd volume in keights binary instead of systemd mount
Systemd mount units cannot do retries in case of failure or dependency
failures, so mounting is being put into the keights binary for better
error handling. Now the of the EBS volume, creating of the filesystem,
and mounting of the filesystem is together in one place instead of having
the mounting be a separate process with more chance of failure. The mount
is now placed directly into /etc/fstab instead of using a systemd unit.