Update Python dependencies
Update all requirements.txt files for Ansible and dependencies. This includes
the requirements.txt used for Packer image builds. Updating Ansible for Packer
requires setting the ansible-remote provisioner setting `use_proxy: false`, to
avoid Ansible hanging. This also requires an updated version of Packer.

The images used for Concourse have been updated to a newer Python version so
that they get a newer version of pip for installing cryptography, which fails
on an older version.