Improve behavior of stack deletion
Deleting the etcd or master auto scaling group could be prone to failure because
the etcd volumes need to be detached from the instances before the stack is
deleted, and sometimes the volumes reattached before the instances were deleted.

This disables health checks and replacement of unhealthy instances in the auto
scaling group so that new machines will not be created that will reattach the
etcd volumes to themselves while the stack is being deleted.