Pass apiserver into instances from CloudFormation
Kube-router needs to communicate with the apiserver on the load balancer when
`replace_kube_proxy: true`. This writes the DNS name of the load balancer to a
file on all instances so it can be mounted as a volume for kube-router. An init
container uses the value from the file to expand a kubeconfig template
placeholder.

This also modifies the kube-router manifest to store the kubeconfig template in a
ConfigMap instead of writing it with a shell heredoc.