Make it easier to build development AMIs
This adds a `dev-mode` variable to the Packer build. When set to `true`,
the Ansible playbook will intentionally exit with an error. This should
be used together with the `-on-error=abort` argument to Packer so that an
image is not created. When using `dev-mode`, the Ansible provisioner's
`keep_inventory_file` option will also be set to `true`, so that Ansible
can be run again manually using the inventory.