Add VPC config to stack lambdas
This ensures lambdas run within the configured VPC when
`keights_stack.lambda_subnets` is defined. The lambdas used by
Keights do not accept any external requests, but this is a
security requirement in some environments.

The security group is configured to allow only outbound traffic
on port 443 so they can access Amazon APIs.

This changes `AWS::IAM::Role` CloudFormation resources to include
`ManagedPolicyArns` instead of the `AWS::IAM::ManagedPolicy`
resources including `Roles`, to ensure that the policy is created
and attached before the Lambdas are deployed.